Add files via upload
A rough demonstration of Discrimination Index (DI) mentioned in article ( I. G. Tekdemir and B. Alboyaci, "A Novel Approach for Improvement of Power Swing Blocking and Deblocking Functions in Distance Relays," in IEEE Transactions on Power Delivery, vol. 32, no. 4, pp. 1986-1994, Aug. 2017. doi: 10.1109/TPWRD.2016.2600638 ) for a 5-cycle synthetically created sine wave. Amplitude of the wave is doubled after 2 cycles, so as to demonstrate a sudden change in system (a fault in a power system, e.g.).